Refactored mixins into a combined file.
Air block drops no pickups.
