+++ src/Blocks/BlockBed.h
@@ -4,7 +4,7 @@
 #pragma once
 
 #include "BlockEntity.h"
-#include "MetaRotator.h"
+#include "Mixins.h"
 #include "ChunkInterface.h"
