+++ src/Blocks/BlockBrewingStand.h
@@ -2,8 +2,7 @@
 #pragma once
 
 #include "BlockEntity.h"
-#include "MetaRotator.h"
-#include "ClearMetaOnDrop.h"
+#include "Mixins.h"
