+++ src/Blocks/BlockChest.h
@@ -4,7 +4,7 @@
 #include "../BlockEntities/ChestEntity.h"
 #include "../BlockArea.h"
 #include "../Entities/Player.h"
-#include "MetaRotator.h"
+#include "Mixins.h"
