+++ src/Blocks/BlockDoor.h
@@ -4,7 +4,7 @@
 #include "BlockHandler.h"
 #include "../Entities/Player.h"
 #include "../Chunk.h"
-#include "MetaRotator.h"
+#include "Mixins.h"
 #include "ChunkInterface.h"
 #include "BlockSlab.h"
