+++ src/Blocks/BlockFurnace.h
@@ -3,7 +3,7 @@
 
 #include "BlockEntity.h"
 #include "../Blocks/BlockPiston.h"
-#include "MetaRotator.h"
+#include "Mixins.h"
