+++ src/Blocks/BlockLever.h
@@ -2,7 +2,7 @@
 
 #include "BlockHandler.h"
 #include "../Chunk.h"
-#include "MetaRotator.h"
+#include "Mixins.h"
 #include "BlockSlab.h"
