+++ src/Blocks/BlockPiston.h
@@ -3,7 +3,7 @@
 
 #include "BlockHandler.h"
 #include <unordered_set>
-#include "ClearMetaOnDrop.h"
+#include "Mixins.h"
 #include "../Item.h"
