+++ src/Blocks/BlockRedstoneRepeater.h
@@ -2,10 +2,9 @@
 #pragma once
 
 #include "BlockHandler.h"
-#include "MetaRotator.h"
+#include "Mixins.h"
 #include "ChunkInterface.h"
 #include "BlockSlab.h"
-#include "ClearMetaOnDrop.h"
 #include "../Chunk.h"
