+++ src/Blocks/BlockTorch.h
@@ -3,7 +3,7 @@
 #include "BlockHandler.h"
 #include "../Chunk.h"
 #include "ChunkInterface.h"
-#include "MetaRotator.h"
+#include "Mixins.h"
