+++ src/Blocks/BlockTrapdoor.h
@@ -2,7 +2,7 @@
 #pragma once
 
 #include "BlockHandler.h"
-#include "MetaRotator.h"
+#include "Mixins.h"
 #include "../EffectID.h"
